refactor: split lean toplevel.rs into a module directory by emission family
Pure mechanical move (3925 lines -> 34-line root + six files): fn_def
(function/mutual-group emission with termination contracts), fuel (fuel
helpers and the floor-window support stacks), lex_list (well-founded
lex-list SCC machinery), type_def (sum/product types and recursive
measures), verify (sample guards, verify blocks, trace proofs), plus
shared rendering helpers. Every external toplevel::X path preserved via
re-exports.

Gates: corpus exports byte-identical to main on both backends; item
parity name-by-name; moved bodies byte-equal modulo fmt rewraps;
proof_spec test list identical.
